Pull out to_nat spec/proof functions
